Moved Navbar styles to own CSS
