+++ agile-blanche/src/app/globals.css
@@ -115,52 +115,3 @@ body {
   }
 }
 
-#nav-icon3 {
-  width: 17px;
-  height: 17px;
-  position: relative;
-  transform: rotate(0deg);
-  transition: cubic-bezier(0.6, -0.28, 0.735, 0.045) s ease-in-out;
-  cursor: pointer;
-}
-
-#nav-icon3 span {
-  display: block;
-  position: absolute;
-  height: 2px;
-  width: 100%;
-  background: #ffffff;
-  border-radius: 3px;
-  opacity: 1;
-  left: 0;
-  transform: rotate(0deg);
-  transition: 0.4s ease-in-out;
-}
-
-#nav-icon3 span:nth-child(1) {
-  top: 0px;
-}
-
-#nav-icon3 span:nth-child(2),
-#nav-icon3 span:nth-child(3) {
-  top: 7px;
-}
-
-#nav-icon3 span:nth-child(4) {
-  top: 14px;
-}
-
-#nav-icon3.open span:nth-child(1),
-#nav-icon3.open span:nth-child(4) {
-  top: 12px;
-  width: 0%;
-  left: 50%;
-}
-
-#nav-icon3.open span:nth-child(2) {
-  transform: rotate(45deg);
-}
-
-#nav-icon3.open span:nth-child(3) {
-  transform: rotate(-45deg);
-}
